additional validation

+++ server/controllers/vehicleController.js
@@ -12,7 +12,7 @@ const getVehicles = asyncHandler(async (req, res) => {
   const filters = req.query;
   let filteredVehicles = vehicles;
 
-  if (JSON.stringify(filters) !== '{}') {
+  if (JSON.stringify(filters) !== "{}") {
     filteredVehicles = vehicles.filter((vehicle) => {
       let isValid = true;
       for (key in filters) {
@@ -57,6 +57,16 @@ const addVehicle = asyncHandler(async (req, res) => {
   const vehicle_type_name = await VehicleType.findById(vehicle_type);
   const fuel_type_name = await FuelType.findById(fuel_type);
 
+  if (!vehicle_type_name) {
+    res.status(400);
+    throw new Error("Vehicle type does not exist");
+  }
+
+  if (!fuel_type_name) {
+    res.status(400);
+    throw new Error("Fuel type does not exist");
+  }
+
   const vehicle = await Vehicle.create({
     ...req.body,
     user: req.user.id,
@@ -91,11 +101,33 @@ const updateVehicle = asyncHandler(async (req, res) => {
     throw new Error("User not authorized");
   }
 
-  const updatedVehicle = await Vehicle.findByIdAndUpdate(
-    req.params.id,
-    req.body,
-    { new: true }
-  );
+  const { vehicle_type, fuel_type } = req.body;
+  let body = { ...req.body };
+
+  if (vehicle_type) {
+    try {
+      const vehicle_type_name = await VehicleType.findById(vehicle_type);
+      body["vehicle_type_name"] = vehicle_type_name.name;
+    } catch (error) {
+      res.status(400);
+      throw new Error("Vehicle type does not exist");
+    }
+  }
+
+  if (fuel_type) {
+    const fuel_type_name = await FuelType.findById(fuel_type);
+
+    if (!fuel_type_name) {
+      res.status(400);
+      throw new Error("Fuel type does not exist");
+    }
+
+    body["fuel_type_name"] = fuel_type_name.name;
+  }
+
+  const updatedVehicle = await Vehicle.findByIdAndUpdate(req.params.id, body, {
+    new: true,
+  });
 
   res.status(200).json(updatedVehicle);
 });
@@ -105,6 +137,7 @@ const updateVehicle = asyncHandler(async (req, res) => {
 // @access Private
 const deleteVehicle = asyncHandler(async (req, res) => {
   const vehicle = await Vehicle.findById(req.params.id);
+
   if (!vehicle || !req.params.id || req.params.id === "") {
     res.status(400);
     throw new Error("Vehicle not found");
@@ -120,6 +153,11 @@ const deleteVehicle = asyncHandler(async (req, res) => {
   if (vehicle.user.toString() !== req.user.id) {
     res.status(401);
     throw new Error("User not authorized");
+  }
+
+  if (vehicle.status === "used") {
+    res.status(400);
+    throw new Error("Vehicle is currently being used");
   }
 
   await vehicle.deleteOne();
